BigfootSightings/requirements.txt
bcrypt==3.2.2
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.2.2
cffi==1.15.0
charset-normalizer==3.3.2
click==8.1.3
colorama==0.4.4
geopy==2.4.1
Flask==2.1.2
Flask-Bcrypt==1.0.1
Flask-Login==0.6.1
Flask-WTF==1.0.1
greenlet==3.0.3
idna==3.7
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.26.4
pandas==1.4.2
psycopg2==2.9.9
pycparser==2.21
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
requests==2.31.0
six==1.16.0
soupsieve==2.5
urllib3==2.2.1
Werkzeug==2.1.2
WTForms==3.0.1
zipp==3.8.0
tqdm==4.66.4

